requirements.txt
google-generativeai>=0.3.0
streamlit>=1.30.0
